pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "CAD"
version = "0.0.1"
description = "Entire Data processing and model training of attendance data"
authors = [
  { name = "Mohan Naredla" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
dependencies = [
    "fastparquet>=2024.11.0",
    "ipykernel>=6.29.5",
    "joblib>=1.5.1",
    "lightgbm>=4.6.0",
    "loguru",
    "matplotlib>=3.10.3",
    "openpyxl>=3.1.5",
    "optuna>=4.4.0",
    "pandas>=2.3.0",
    "pip",
    "pyarrow>=20.0.0",
    "python-dotenv",
    "ruff",
    "scikit-learn>=1.7.0",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "shap>=0.48.0",
    "tqdm",
    "typer",
]
requires-python = "~=3.13.0"


[tool.ruff]
line-length = 99
src = ["CAD"]
include = ["pyproject.toml", "CAD/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["CAD"]
force-sort-within-sections = true

